# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "endPoind ",
        "version": "1.0"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/liste_candidature": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Obtenir la liste de toutes les candidatures",
                "operationId": "d18779ea2660d6787b277d6833993e12",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code_valide": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "La liste des candidatures a été bien récupérée."
                                        },
                                        "liste des candidatures": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FormationUtilisateur"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "La liste des candidatures est indisponible",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/liste_candidat_refus": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Obtenir la liste des candidatures refusées",
                "operationId": "e8e3ad0e4a818290e277f556ec46936e",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code_valide": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "La liste des candidatures refusées a été bien récupérée."
                                        },
                                        "liste des candidatures refusées": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FormationUtilisateur"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "la liste des candidatures refusées est indisponible",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/liste_candidat_accept": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Obtenir la liste des candidatures acceptées",
                "operationId": "412ccba2854c199fe0a4a6d054338e3b",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code_valide": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "La liste des candidatures acceptées a été bien récupérée."
                                        },
                                        "liste des candidatures acceptées": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FormationUtilisateur"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "la liste des candidatures acceptées est indisponible",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/enregistrer_candidature/{id}": {
            "post": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Enregistrer une nouvelle candidature",
                "operationId": "07686ae6ac5624bde14e793ca1e59b66",
                "requestBody": {
                    "description": "Informations de la candidature",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_formation"
                                ],
                                "properties": {
                                    "id_formation": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidature enregistré avec Succè",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidature enregistrée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "échec de la candidature",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/accepter_candidature/{id}": {
            "post": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Accepter une candidature",
                "description": "Accepte la candidature associée à l'ID spécifié.",
                "operationId": "75fcf11d657489d7540221ab58ee112d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "statut_candidature": {
                                        "description": "Nouveau statut de la candidature (Accepter)",
                                        "type": "string",
                                        "example": "Accepter"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidature acceptée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidature acceptée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidature non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidature non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/candidature/{id}/refuser": {
            "post": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Refuser une candidature",
                "description": "Refuse la candidature associée à l'ID spécifié.",
                "operationId": "5946263b45e3d54d923c25663429ba87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "statut_candidature": {
                                        "description": "Nouveau statut de la candidature (Refuser)",
                                        "type": "string",
                                        "example": "Refuser"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidature refusée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidature refusée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidature non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidature non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}